Move the setup to our own fork of just-environment-puppeteer
